AG-12847 - extract feature modules (#8794)
* AG-12847 - extract column autosize module

* AG-12847 - simplify column key creator

* AG-12847 - extract column autosize module

* AG-12847 - column move module

* AG-12847 - extract agg column name logic

* AG-12847 - controls column module

* AG-12847 - remove unnecessary getter methods

* AG-12847 - remove unnecessary getter methods

* AG-12847 - column resize module

* AG-12847 - remove unnecessary getter methods

* AG-12847 - remove column event dispatcher

* AG-12847 - drag and drop module

* AG-12847 - drag and drop image comp

* AG-12847 - type export services

* AG-12847 - extract pinned row module

* AG-12847 - pull out api modules

* AG-12847 - extract cell range feature

* AG-12847 - extract sticky row feature

* AG-12847 - extract drag interfaces

* AG-12847 - extract selection

* AG-12847 - rearrange modules

* AG-12847 - extract selection

* AG-12847 - extract sorting

* AG-12847 - extract expansion

* AG-12847 - extract value cache

* AG-12847 - extract expressions

* AG-12847 - extract user components

* AG-12847 - fix all module imports/exports

* AG-12847 - fix tests

* AG-12847 - extract position utils

* AG-12847 - extract ag component utils

+++ packages/ag-grid-enterprise/src/serverSideRowModel/transactionManager.ts
@@ -19,9 +19,9 @@ interface AsyncTransactionWrapper {
 export class TransactionManager extends BeanStub implements NamedBean, IServerSideTransactionManager {
     beanName = 'ssrmTransactionManager' as const;
 
-    private valueCache: ValueCache;
+    private valueCache?: ValueCache;
     private serverSideRowModel: ServerSideRowModel;
-    private selectionService: ServerSideSelectionService;
+    private selectionService?: ServerSideSelectionService;
 
     public wireBeans(beans: BeanCollection): void {
         this.valueCache = beans.valueCache;
@@ -109,7 +109,7 @@ export class TransactionManager extends BeanStub implements NamedBean, IServerSi
         this.asyncTransactions = transactionsToRetry;
 
         if (atLeastOneTransactionApplied) {
-            this.valueCache.onDataChanged();
+            this.valueCache?.onDataChanged();
             this.eventService.dispatchEvent({ type: 'storeUpdated' });
         }
 
@@ -139,8 +139,8 @@ export class TransactionManager extends BeanStub implements NamedBean, IServerSi
         if (!hasStarted) {
             return { status: ServerSideTransactionResultStatus.StoreNotStarted };
         } else if (res) {
-            this.valueCache.onDataChanged();
-            if (res.remove) {
+            this.valueCache?.onDataChanged();
+            if (res.remove && this.selectionService) {
                 const removedRowIds = res.remove.map((row) => row.id!);
                 this.selectionService.deleteSelectionStateFromParent(transaction.route || [], removedRowIds);
             }
